+++ social_media_catalog/__init__.py
@@ -3,6 +3,5 @@
 # SPDX-License-Identifier: BSD-3-Clause
 """Export all functions and modules for social_media_catalog."""
 
-import social_media_catalog.youtube
 
-__all__ = [social_media_catalog.youtube]
+__all__ = ["social_media_catalog.youtube"]