+++ social_media_catalog/youtube/playlist.py
@@ -3,28 +3,25 @@
 # SPDX-License-Identifier: BSD-3-Clause
 """A template to store the list of videos from a playlist into json."""
 
-from itertools import zip_longest
+from dataclasses import dataclass, asdict
 from typing import Any
-from typing_extensions import Self
 
 from social_media_catalog.youtube import Video
 
 
+@dataclass(frozen=True, order=True)
 class Playlist:
     """A class to store the videos with the playlist name and id."""
 
-    def __init__(
-        self: Self, playlist_id: str, playlist_name: str, videos: list[Video]
-    ) -> None:
-        self.playlist_id: str = playlist_id
-        self.playlist_name: str = playlist_name
-        self.videos: list[Video] = videos
+    playlist_id: str
+    playlist_name: str
+    videos: list[Video]
 
-    def to_dict(self: Self) -> dict[str, Any]:
+    def to_dict(self) -> dict[str, Any]:
         """Convert the VideoList object to a dictionary for json usage.
 
         Args:
-            self (Self): The object instance that stores the
+            self: The object instance that stores the
             videos in a schema.
 
         Returns:
@@ -33,17 +30,16 @@ class Playlist:
         videos_list = []
 
         for video in self.videos:
-            videos_list.append(video.to_dict())
+            videos_list.append(asdict(video))
 
         return {
-            "$Schema": self.schema,
             "PlaylistId": self.playlist_id,
             "PlaylistName": self.playlist_name,
             "Videos": videos_list,
         }
 
     @classmethod
-    def from_json(cls: Self, data: dict[str, Any], /) -> Self:
+    def from_json(cls, data: dict[str, Any], /):
         """Convert the formatted json data back to a VideoList object.
 
         Args:
@@ -55,7 +51,7 @@ class Playlist:
             keys: 'PlaylistId', 'PlaylistName' or 'Videos'.
 
         Returns:
-            VideoList: The object that was converted from a
+            Self: The object that was converted from a
             dictionary from json.
         """
 
@@ -75,11 +71,3 @@ class Playlist:
             "following keys:\n"
             "'PlaylistId', 'PlaylistName' or 'Videos'"
         )
-
-    def __eq__(self: Self, other: object) -> bool:
-        if not isinstance(other, Self):
-            return False
-        for a, b in zip_longest(self.videos, other.videos):
-            if not a == b:
-                return False
-        return True