+++ tests/__init__.py
@@ -4,4 +4,4 @@
 """Import module for coverage and unit tests for the 'video-lister'
 package."""
 
-import unittest
+import pytest